+++ src/backend/nsl/nsl_geom_linesim.h
@@ -61,15 +61,15 @@ double nsl_geom_linesim_positional_squared_error(const double xdata[], const dou
 */
 double nsl_geom_linesim_area_error(const double xdata[], const double ydata[], const size_t n, const size_t index[]);
 
-/* calculates tolerance by using diagonal distance of all data point clip area 
+/* calculates tolerance by using diagonal distance of all data point clip area
 	divided by n */
 double nsl_geom_linesim_clip_diag_perpoint(const double xdata[], const double ydata[], const size_t n);
 
-/* calculates tolerance from clip area 
+/* calculates tolerance from clip area
 	divided by n */
 double nsl_geom_linesim_clip_area_perpoint(const double xdata[], const double ydata[], const size_t n);
 
-/* calculates tolerance from average distance of following point 
+/* calculates tolerance from average distance of following point
 	divided by n */
 double nsl_geom_linesim_avg_dist_perpoint(const double xdata[], const double ydata[], const size_t n);
 
@@ -101,7 +101,7 @@ double nsl_geom_linesim_douglas_peucker_variant(const double xdata[], const doub
 	index: index of reduced points
 	-> returns final number of points
 */
-size_t nsl_geom_linesim_nthpoint(const size_t n, const size_t step, size_t index[]);
+size_t nsl_geom_linesim_nthpoint(const size_t n, const int step, size_t index[]);
 
 /* radial distance line simplification
 	xdata, ydata: data points
